extinfo time reporting fix
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk@3534 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ ac/source/src/serverms.cpp
@@ -207,7 +207,7 @@ void serverms(int mode, int numplayers, int minremain, char *smapname, int milli
             {
                 case EXT_UPTIME:        // uptime in seconds
                 {
-                    putint(po, millis/1000);
+                    putint(po, uint(millis)/1000);
                     break;
                 }
